services/verticals/papabase/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.0
google-cloud-firestore==2.16.1
httpx==0.27.0
pydantic==2.7.1
fastapi==0.111.0
